requirements.txt
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
SpeechRecognition==3.10.4
gTTS==2.5.3
google-generativeai==0.8.3
openai==1.51.2
numpy>=1.26.4
